requirements.txt
python-telegram-bot==20.5
google-generativeai==0.5.4